Embedded chart widget into chart page, added lazy loading of ECharts.
