+++ bundles/org.openhab.ui/web/src/components/widgets/system/oh-chart.vue
@@ -1,29 +1,16 @@
 <template>
-  <div class="oh-chart-page-container" v-bind="config" :style="{ height: config.height || '300px' }" >
-    <oh-chart-page :context="context" />
-  </div>
+  <oh-chart-component :context="this.context" />
 </template>
-
-<style lang="stylus">
-.oh-chart-page-container
-  position relative
-  overflow hidden
-  width 100%
-
-.oh-chart-page-container .oh-chart-page-chart
-  background-color transparent !important
-</style>
 
 <script>
 import mixin from '../widget-mixin'
-import OhChartPage from '../chart/oh-chart-page.vue'
-import { OhChartPageDefinition } from '@/assets/definitions/widgets/chart/page'
+import { OhChartDefinition } from '@/assets/definitions/widgets/system'
 
 export default {
   mixins: [mixin],
   components: {
-    OhChartPage
+    'oh-chart-component': () => import(/* webpackChunkName: "oh-chart-component" */ './oh-chart-component.vue')
   },
-  widget: OhChartPageDefinition
+  widget: OhChartDefinition
 }
 </script>
